refactor: move rubrics models into models subfolder
